Fix build, tests, examples for SDL3 as of 2023-07-31 (#5)
* Check /usr/local/lib for libSDL on Linux

* Fix build, tests, examples for SDL3 as of 2023-07-30

This commit fixes numerous build and test failures that pop up when
linking against an updated SDL3-sys that uses recent a recent SDL3
git version (fb68e8464638976e13c6a31eb8d3a7884e4b833e used here).

Note that this version depends on an open PR to sdl3-sys-rs (to
allow for testing).

New feature:
- Feature `test-mode` allows running the integration tests without
  failing the Sdl::init() sanity check; the documentation reflects
  this now.

Missing:
- examples/audio-wav.rs is currently disabled; it needs to be ported
  from AudioCVT (which has been retired) to AudioStream.
- The examples need testing.
- The docs still refer to sdl2 all over the place.

+++ examples/animation.rs
@@ -75,14 +75,14 @@ fn main() -> Result<(), String> {
         let ticks = timer.ticks() as i32;
 
         // set the current frame for time
-        source_rect_0.set_x(32 * ((ticks / 100) % frames_per_anim));
-        dest_rect_0.set_x(1 * ((ticks / 14) % 768) - 128);
+        source_rect_0.set_x((32 * ((ticks / 100) % frames_per_anim)) as f32);
+        dest_rect_0.set_x((1 * ((ticks / 14) % 768) - 128) as f32);
 
-        source_rect_1.set_x(32 * ((ticks / 100) % frames_per_anim));
-        dest_rect_1.set_x((1 * ((ticks / 12) % 768) - 672) * -1);
+        source_rect_1.set_x((32 * ((ticks / 100) % frames_per_anim)) as f32);
+        dest_rect_1.set_x(((1 * ((ticks / 12) % 768) - 672) * -1) as f32);
 
-        source_rect_2.set_x(32 * ((ticks / 100) % frames_per_anim));
-        dest_rect_2.set_x(1 * ((ticks / 10) % 768) - 128);
+        source_rect_2.set_x((32 * ((ticks / 100) % frames_per_anim)) as f32);
+        dest_rect_2.set_x((1 * ((ticks / 10) % 768) - 128) as f32);
 
         canvas.clear();
         // copy the frame to the canvas
