+++ src/sdl3/rwops.rs
@@ -1,6 +1,6 @@
 use crate::get_error;
 use libc::c_void;
-use libc::{c_char, c_int};
+use libc::c_char;
 use std::ffi::CString;
 use std::io;
 use std::marker::PhantomData;
@@ -58,7 +58,7 @@ impl<'a> RWops<'a> {
     #[doc(alias = "SDL_RWFromConstMem")]
     pub fn from_bytes(buf: &'a [u8]) -> Result<RWops<'a>, String> {
         let raw =
-            unsafe { sys::SDL_RWFromConstMem(buf.as_ptr() as *const c_void, buf.len() as c_int) };
+            unsafe { sys::SDL_RWFromConstMem(buf.as_ptr() as *const c_void, buf.len()) };
 
         if raw.is_null() {
             Err(get_error())
@@ -92,7 +92,7 @@ impl<'a> RWops<'a> {
     /// This method can only fail if the buffer size is zero.
     #[doc(alias = "SDL_RWFromMem")]
     pub fn from_bytes_mut(buf: &'a mut [u8]) -> Result<RWops<'a>, String> {
-        let raw = unsafe { sys::SDL_RWFromMem(buf.as_ptr() as *mut c_void, buf.len() as c_int) };
+        let raw = unsafe { sys::SDL_RWFromMem(buf.as_ptr() as *mut c_void, buf.len()) };
 
         if raw.is_null() {
             Err(get_error())
